[FE] ♻️ NewChat, Chat 겹치는 세부구현 hook으로 정리
