+++ client/src/components/inquiry/NewChat.tsx
@@ -1,17 +1,13 @@
 'use client';
 
 import { useRouter } from 'next/navigation';
-import { useEffect, useRef, useState } from 'react';
-
-import { CompatClient, Stomp, StompSubscription } from '@stomp/stompjs';
-import SockJS from 'sockjs-client';
 
 import useUserStore from '@/stores/userStore';
 import useChatStore from '@/stores/chatStore';
 
-import { ChatInput, ChatBox } from '.';
+import useNewChatAndExistChatConnect from '@/hooks/useNewChatAndExistChatConnect';
 
-import { Chat } from '@/types/data';
+import { ChatInput, ChatBox } from '.';
 
 import checkForToken from '@/utils/checkForToken';
 
@@ -20,33 +16,15 @@ interface NewChatProps {
 }
 
 export default function NewChat({ role }: NewChatProps) {
-  const client = useRef<CompatClient>();
-  const scrollRef = useRef<HTMLDivElement | null>(null);
-
-  const [connected, setConnected] = useState(false);
-  const [chat, setChat] = useState<Chat[]>([]);
-
   const router = useRouter();
 
-  const { message, setMessage, roomId } = useChatStore();
-  const { accessToken, refreshToken, displayName, userId, setClear } =
-    useUserStore();
+  const { message, roomId, isNewChatConnect, setMessage } = useChatStore();
+  const { displayName, userId, setClear } = useUserStore();
+
+  const { setConnected, client, scrollRef, chat, connected } =
+    useNewChatAndExistChatConnect(isNewChatConnect);
 
   const { authVerify } = checkForToken();
-
-  const url = process.env.NEXT_PUBLIC_API_URL;
-
-  let subscription: StompSubscription | undefined;
-
-  const entryMessage = () => {
-    const adminId = 101;
-
-    client?.current?.send(
-      `/pub/chatRoom/enter`,
-      {},
-      JSON.stringify({ senderId: +userId, chatRoomId: +roomId, adminId }),
-    );
-  };
 
   const newMessge = {
     senderId: +userId,
@@ -62,9 +40,7 @@ export default function NewChat({ role }: NewChatProps) {
       authVerify() === 'Refresh Token Expired'
     ) {
       return (
-        alert(
-          '토큰이 만료되었습니다. 로그아웃 후 다시 로그인 해주시길 바랍니다.',
-        ),
+        alert('토큰이 만료되었습니다. 다시 로그인 해주시길 바랍니다.'),
         setConnected(false),
         setClear(),
         setMessage(''),
@@ -73,47 +49,9 @@ export default function NewChat({ role }: NewChatProps) {
     }
 
     client?.current?.send(`/pub/chatRoom/send`, {}, JSON.stringify(newMessge));
+
     setMessage('');
   };
-
-  useEffect(() => {
-    if (roomId) {
-      client.current = Stomp.over(() => new SockJS(`${url}/wss`));
-      client.current.debug = () => {};
-      client.current.connect(
-        {
-          Authorization: accessToken,
-          refresh: refreshToken,
-        },
-        () => {
-          subscription = client?.current?.subscribe(
-            `/sub/chatRoom/${roomId}`,
-            (payload) => {
-              const receivedMessage: Chat = JSON.parse(payload.body);
-
-              setChat((previousChat) => [...previousChat, receivedMessage]);
-            },
-          );
-
-          entryMessage();
-          setConnected(true);
-        },
-      );
-    }
-
-    return () => {
-      client.current?.disconnect(() => {
-        subscription?.unsubscribe();
-        setConnected(false);
-      });
-    };
-  }, [roomId]);
-
-  useEffect(() => {
-    if (!scrollRef.current) return;
-
-    scrollRef.current.scrollTop = scrollRef.current.scrollHeight;
-  }, [chat]);
 
   return (
     <section className="w-full">
@@ -139,5 +77,5 @@ export default function NewChat({ role }: NewChatProps) {
 
 const CHAT_STYLE = {
   user: 'w-[220px] h-[232px] mb-[7px]',
-  admin: 'lg:w-[392px] sm:w-[312px] h-[327px]',
+  admin: 'w-full h-[327px]',
 };
